% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAIT.R
\name{BAIT}
\alias{BAIT}
\title{BAIT -- master function to process strand-seq libraries into BAIT ideograms}
\usage{
BAIT(path = ".", splitBy = 2e+05, readQual = 10, pairedEnd = TRUE,
  plotBy = "lib", plotName = NULL, chroms = "all", verbose = TRUE)
}
\arguments{
\item{path}{String denoting location of Strand-seq bam files (default is ".")}

\item{splitBy}{numeric value of binsize for plotting. default is 200000}

\item{readQual}{Integer dictating the minimal mapping quality required for a read to be accepted. Default is 10.}

\item{pairedEnd}{Whether the bam files being read are in paired end format. Default is TRUE. Note,}

\item{plotBy}{Whether to plot by library ('lib') or chromosome ('chr')}

\item{plotName}{character which determines file name to be saved. Default is to open an R plot from the terminal}

\item{chroms}{vector of chromosome number to prevent contig plotting. eg for humans use 1:24. Default is 'all'}

\item{verbose}{prints messages to the terminal (default is TRUE)}
}
\value{
ordered contigs in bed format. Depending on options, intermediate files and plots will also be generated
}
\description{
BAIT -- master function to process strand-seq libraries into BAIT ideograms
}
\examples{
#Get a list of BAM files containing libraries for cells from the same organism, aligned to the same genome
#In this case these are the example BAM files provided with the package (hence the call to system.file);
data("exampleDividedChr")

example.dir <- file.path(system.file(package='contiBAIT'), 'extdata')

orderedContigs <- contiBAIT(path=example.dir, cluster=1, clusNum=1, filter=exampleDividedChr, pairedEnd=FALSE)
}

