% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotLinkageGroupCalls.R
\docType{methods}
\name{barplotLinkageGroupCalls,LinkageGroupList-method}
\alias{barplotLinkageGroupCalls}
\alias{barplotLinkageGroupCalls,LinkageGroupList,LinkageGroupList-method}
\alias{barplotLinkageGroupCalls,LinkageGroupList-method}
\title{Bar plot all linkage groups, with the true chromosomes of contigs coloured.}
\usage{
\S4method{barplotLinkageGroupCalls}{LinkageGroupList}(object, assemblyBED,
  by = "lg", returnTable = FALSE)
}
\arguments{
\item{object}{LinkageGroupList, as generated by clusterContigs}

\item{assemblyBED}{table from a BED containing assembly information about the contigs, including length and chromosome
Note that the rownames of assemblyBED should be the contig names, as they are used in object, and the first column (chromosome name) will used to order by chromosome if 'chr' option used in by parameter. 
To use a bam file header, the product of makeChrTable(bamFile), or divideMyChr with or without the asBed option is suitable for input}

\item{by}{whether to plot by linkage group (if 'lg') or chromosomes ('chr')}

\item{returnTable}{TRUE to return chromosome length matrix.
Note to include legend, use legend=rownames(chr.table) for by='lg', and legend=colnames(chr.table) for by='chr'}
}
\value{
a matrix of lengths of each chromosome (rows) in each linkage group (columns)
}
\description{
Bar plot all linkage groups, with the true chromosomes of contigs coloured.
}
\examples{
data("exampleLGList")
data("exampleDividedChr")

barplotLinkageGroupCalls(exampleLGList, exampleDividedChr)

}

