% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotLinkageGroupCalls.R
\name{barplotLinkageGroupCalls}
\alias{barplotLinkageGroupCalls}
\title{Bar plot all linkage groups, with the true chromosomes of contigs coloured.}
\usage{
barplotLinkageGroupCalls(linkageGroups, assemblyBED, by = "lg",
  returnTable = FALSE, saveFile = FALSE, ...)
}
\arguments{
\item{linkageGroups}{list of vectors, each specifying which contigs belong in which linkage group}

\item{assemblyBED}{table from a BED containing assembly information about the contigs, including length and chromosome
Note that the rownames of assemblyBED should be the contig names, as they are used in linkageGroups. 
To use a bam file header, the product of makeChrTable(bamFile) with or without the asBed option is suitable for input}

\item{returnTable}{TRUE to return chromosome length matrix.}

\item{saveFile}{string providing a file name to save the plot as a pdf (with legend). Default is no saved file.}

\item{...}{any additional parameters to pass down to barplot()
Note to include legend, use legend=rownames(chr.table) for by='lg', and legend=colnames(chr.table) for by='chr'}

\item{by='lg'}{whether to plot by linkage group (if 'lg') or chromosomes ('chr')}
}
\value{
a matrix of lengths of each chromosome (rows) in each linkage group (columns)
}
\description{
Bar plot all linkage groups, with the true chromosomes of contigs coloured.
}

