% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ChrTable-class}
\alias{ChrTable-class}
\title{A class for storing chromosome/fragment lengths}
\description{
\describe{
 This class is data.frame of two or three vectors. The first is a character vector representing the chromosome name, the second is an integer of either 
 length of the molecule or the start of the molecule, and the (optional) third is a integer of the end of the molecule.
 Note the rowname represents the fragment name.
}
}

