% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineZeroDistContigs.R
\name{combineZeroDistContigs}
\alias{combineZeroDistContigs}
\title{Function to combine contigs that are right next to each other.}
\usage{
combineZeroDistContigs(linkageStrands, rawStrandTable, lg)
}
\arguments{
\item{linkageStrands}{correctly oriented strandTable, but only containing the rows for this linkage group}
}
\value{
a two-member list, first=new strand table, second=list mapping new merged contigs to old
}
\description{
Function to combine contigs that are right next to each other.
}

