% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideMyChr.R
\name{divideMyChr}
\alias{divideMyChr}
\title{Creates a data.frame of chromosomes split into user-derived bin sizes}
\usage{
divideMyChr(chrTable, splitBy = 2e+05)
}
\arguments{
\item{chrTable}{data.frame consisting of either chromosome name and length, or chromosome name, start and end. Can
generate from makeChrTable}

\item{splitBy}{numeric value dictating the size of each bin in nucleotides. Default is 200000.}
}
\description{
Creates a data.frame of chromosomes split into user-derived bin sizes
}
\examples{
#Get an example BAM file and generate a chromosome table featuring fragment names and lengths, then split into 50kb fragments

data("exampleChrTable")
dividedChr <- divideMyChr(exampleChrTable, splitBy=1000000)
}

