% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleChrTable}
\alias{exampleChrTable}
\title{Example of a ChromosomeTable, containing contigs and their lengths}
\description{
Example of a ChromosomeTable, containing contigs and their lengths
}
\keyword{data}
