% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleContigOrder}
\alias{exampleContigOrder}
\title{Example of a ContigOrdering table, containing a list with a matrix of ordered groups element
and a StrandStateList element}
\description{
Example of a ContigOrdering table, containing a list with a matrix of ordered groups element
and a StrandStateList element
}
\keyword{data}
