% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleCrickFreq}
\alias{exampleCrickFreq}
\title{An example Crick strand frequences extracted from BAMS by strandSeqFreqTable where BAITtables=TRUE}
\description{
An example Crick strand frequences extracted from BAMS by strandSeqFreqTable where BAITtables=TRUE
}
\keyword{data}

