% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleLGList}
\alias{exampleLGList}
\title{Example of a LinkageGroupList output from clusterContigs}
\description{
Example of a LinkageGroupList output from clusterContigs
}
\keyword{data}
