% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleLibList}
\alias{exampleLibList}
\title{Example of a LibraryGroupList, containing library names}
\description{
Example of a LibraryGroupList, containing library names
}
\keyword{data}
