% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleReadCounts}
\alias{exampleReadCounts}
\title{Example of read counts extracted from BAMS by strandSeqFreqTable}
\description{
Example of read counts extracted from BAMS by strandSeqFreqTable
}
\keyword{data}

