% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleStrandFrequencies}
\alias{exampleStrandFrequencies}
\title{Example of strand frequencies extracted from BAMS by strandSeqFreqTable}
\description{
Example of strand frequencies extracted from BAMS by strandSeqFreqTable
}
\keyword{data}

