% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleWatsonFreq}
\alias{exampleWatsonFreq}
\title{An example Watson strand frequences extracted from BAMS by strandSeqFreqTable where BAITtables=TRUE}
\description{
An example Watson strand frequences extracted from BAMS by strandSeqFreqTable where BAITtables=TRUE
}
\keyword{data}

