% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleWCMatrix}
\alias{exampleWCMatrix}
\title{An example StrandStateMatrix containing WW, CC and WC calls for contigs}
\description{
An example StrandStateMatrix containing WW, CC and WC calls for contigs
}
\keyword{data}

