% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractReads.R
\name{extractReads}
\alias{extractReads}
\title{extractReads -- function to make bed file from all reads occupying uncalled regions from createOrientationBed}
\usage{
extractReads(locations, uncalled)
}
\arguments{
\item{locations}{data.frame of strand frequencies created by createReadLocations}

\item{uncalled}{data.frame of uncalled regions created createOrientationBed}
}
\value{
data.frame in bed format of all uncalled reads
}
\description{
extractReads -- function to make bed file from all reads occupying uncalled regions from createOrientationBed
}

