% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSexGroups.R
\docType{methods}
\name{findSexGroups,LinkageGroupList,StrandStateMatrix-method}
\alias{findSexGroups}
\alias{findSexGroups,LinkageGroupList,StrandStateMatrix-method}
\title{Bar plot all linkage groups, with the true chromosomes of contigs coloured.}
\usage{
\S4method{findSexGroups}{LinkageGroupList,StrandStateMatrix}(linkageGroupList,
  strandStateMatrix, callThreshold = 0.2)
}
\arguments{
\item{linkageGroupList}{a LinkageGroupList object as generated by clusterContigs}

\item{strandStateMatrix}{a StrandStateMatrix object as generated by preprocessStrandTable}

\item{callThreshold}{a value between 0 and 1 to threshold the proportion of libraries in which scaffolds are heterozygous (ie
unlikely to be a monosome sex chromosome). Default value is 0.2}
}
\value{
the same LinkageGroupList entered into the function, but with potential sex chromosomes highlighted in the name attribute
}
\description{
Bar plot all linkage groups, with the true chromosomes of contigs coloured.
}

