% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipOrderedLinkageGroups.R
\docType{methods}
\name{flipOrderedLinkageGroups,ContigOrdering,StrandStateList,LinkageGroupList,StrandStateMatrix-method}
\alias{flipOrderedLinkageGroups,ContigOrdering,StrandStateList,LinkageGroupList,StrandStateMatrix-method}
\alias{flipOrderedLinkageGroups}
\alias{flipOrderedLinkageGroups-ContigOrdering-StrandStateList-method}
\title{flipOrderedLinkageGroups -- Simple greedy algorithm to find misorientations within previously ordered LG.}
\usage{

  \S4method{flipOrderedLinkageGroups}{ContigOrdering,StrandStateList,LinkageGroupList,StrandStateMatrix}(contigOrdering,
  orderFrame, linkageGroupList, strandStateMatrix, whichLG = NULL,
  maxiter = 100, dissimilarityCutoff = 0.3)
}
\arguments{
\item{contigOrdering}{a data.frame of ordered contigs with linkage group names of class ContigOrdering}

\item{orderFrame}{a list of StrandStateMatrix elements of class StrandStateList}

\item{linkageGroupList}{List of vectors containing names of contigs belonging to each LG of type LinkageGroupList. Should be ordered as the output of orderAllLinkageGroups[[3]]}

\item{strandStateMatrix}{a StrandStateMatrix to be reoriented}

\item{whichLG}{vector of integers specifying the element(s) of linkageGroupList to be ordered (i.e. which specific linkage groups to try to order). Default is all LGs.}

\item{maxiter}{The number of iterations used to find the optimal path. Default is 100}

\item{dissimilarityCutoff}{The minimum amount of dissimilarity needed to start reorienting contigs. Default is 0.3.}
}
\value{
a StrandStateMatrix with misorientations resolved
}
\description{
flipOrderedLinkageGroups -- Simple greedy algorithm to find misorientations within previously ordered LG.
}
