% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatPlotLinkageGroupCalls.R
\name{heatPlotLinkageGroupCalls}
\alias{heatPlotLinkageGroupCalls}
\title{Heat plot linkage groups against true chromosomes, by lengths in MB}
\usage{
heatPlotLinkageGroupCalls(linkageGroups, assemblyBED, ...)
}
\arguments{
\item{linkageGroups}{list of vectors, each specifying which contigs belong in which linkage group}

\item{assemblyBED}{table from a BED containing assembly information about the contigs, including length and chromosome
Note that the rownames of assemblyBED should be the contig names, as they are used in linkageGroups}

\item{...}{any additional parameters to pass down to heatmap.2()}
}
\description{
Heat plot linkage groups against true chromosomes, by lengths in MB
}

