% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightAssemblyErrors.R
\name{highlightAssemblyErrors}
\alias{highlightAssemblyErrors}
\alias{highlightAssemblyErrors,highlightAssemblyErrors-GRanges-method}
\title{highlightAssemblyErrors -- Master function to identify misorientations and chimeras in the assembly}
\usage{
highlightAssemblyErrors(path, splitBy = 1e+06, cluster = 1,
  clusterParam = NULL, pairedEnd = TRUE, qual = 10, gapFile = NULL,
  verbose = TRUE)
}
\arguments{
\item{path}{String denoting location of Strand-seq bam files}

\item{splitBy}{integer determining the average size contigs should be split by}

\item{cluster}{Number of times to recluster and take the consensus of. If NULL, clustering is 
run only once.}

\item{clusterParam}{optional \code{BiocParallelParam} specifying cluster to use for parallel execution.
When \code{NULL}, execution will be serial.}

\item{pairedEnd}{Whether the bam files being read are in paired end format. Default is TRUE. Note,}

\item{qual}{Mapping quality threshold. Default is 1}

\item{gapFile}{A GRanges object consisting of start and end locations of assembly gaps (defaul it NULL)}

\item{verbose}{prints messages to the terminal (default is TRUE)}
}
\value{
a directional ChrTable object that can be used in downstream functions (strandSeqFreqTable)
}
\description{
highlightAssemblyErrors -- Master function to identify misorientations and chimeras in the assembly
}

