% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogramPlot.R
\docType{methods}
\name{ideogramPlot,StrandReadMatrix,StrandReadMatrix,ChrTable-method}
\alias{ideogramPlot}
\alias{ideogramPlot,StrandReadMatrix,StrandReadMatrix,ChrTable-method}
\alias{ideogramPlot,StrandReadMatrix,StrandReadMatrix-method,ChrTable,ChrTable-method}
\title{ideogramPlot -- plots BAIT-like ideograms}
\usage{

  \S4method{ideogramPlot}{StrandReadMatrix,StrandReadMatrix,ChrTable}(WatsonFreqList,
  CrickFreqList, chrTable, plotBy = "lib", sizeProportional = NULL,
  showPage = NULL, orderFrame = NULL, orientationData = NULL,
  verbose = TRUE)
}
\arguments{
\item{WatsonFreqList}{data.frame of Watson calls. Product of strandSeqFreqTable[[3]] when BAITtables=TRUE}

\item{CrickFreqList}{data.frame of Crick calls. Product of strandSeqFreqTable[[4]] when BAITtables=TRUE}

\item{chrTable}{A data.frame consisting of chromosomes and lengths. Generated by makeChrTable(). Note rownames equal to chromosome names are required}

\item{plotBy}{Whether to generate a plot for each library ('lib') or a plot for each chromosome ('chr')}

\item{sizeProportional}{The plotter will divide each fragment into a discrete number of bins of size sizeProportional and distribute reads across them. For example, when set to 200000, the
all contigs in each linkage group will be split into 200kb fragments and each fragment will have an equal proportion of the reads derived from that fragment. This ultimately rescales the ideograms
such that LGs with the greatest DNA content are the biggest, as opposed to teh LGs with the most contigs. Default value is NULL}

\item{showPage}{Integer specifying which LG (if plotBy='chr') or libraries (if plotBy='lib') to plot. Useful when not plotting to a file, or when wishing to subset data. Default is NULL}

\item{orderFrame}{ordered data.frame of contigs (produced by orderAllLinkageGroups). Default is FALSE, where plots will be made from elements in chrTable.}

\item{orientationData}{ChrTable of contig orientations telling which reads to flip Watson and Crick counts}

\item{verbose}{prints messages to the terminal (default is TRUE)}
}
\value{
ordered contigs in bed format. Depending on options, intermediate files and plots will also be generated
}
\description{
ideogramPlot -- plots BAIT-like ideograms
}
\examples{
data("exampleWatsonFreq")
data("exampleCrickFreq")
data('exampleDividedChr')

singleWatsonLibrary <- StrandReadMatrix(exampleWatsonFreq[,2, drop=FALSE])
singleCrickLibrary <- StrandReadMatrix(exampleCrickFreq[,2, drop=FALSE])

ideogramPlot(singleWatsonLibrary, singleCrickLibrary, exampleDividedChr)
}

