% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBoxPlot.R
\docType{methods}
\name{makeBoxPlot,ChrTable,LinkageGroupList-method}
\alias{makeBoxPlot}
\alias{makeBoxPlot,ChrTable,LinkageGroupList-method}
\alias{makeBoxPlot,LinkageGroupList,LinkageGroupList-method}
\title{makeBoxPlot -- creates boxplot of contigs included in the analysis vs those excluded}
\usage{
\S4method{makeBoxPlot}{ChrTable,LinkageGroupList}(chrTable, linkage.contigs,
  sex.contigs = NULL)
}
\arguments{
\item{chrTable}{A data.frame consisting of chromosomes and lengths. Generated by makeChrTable(). Note rownames equal to chromosome names are required}

\item{linkage.contigs}{A list of clustered contigs, generated by clusterContigs()}

\item{sex.contigs}{vector of contigs that appear to be monosomic (ie belonging to sex chromosomes)}
}
\value{
an box plot of included and excluded contigs
}
\description{
makeBoxPlot -- creates boxplot of contigs included in the analysis vs those excluded
}

