% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBoxPlot.R
\name{makeBoxPlot}
\alias{makeBoxPlot}
\title{makeBoxPlot -- creates boxplot of contigs included in the analysis vs those excluded}
\usage{
makeBoxPlot(chrTable, linkage.contigs, sex.contigs = 1, saveFile = FALSE)
}
\arguments{
\item{chrTable}{A data.frame consisting of chromosomes and lengths. Generated by makeChrTable()}

\item{linkage.contigs}{A list of clustered contigs, generated by clusterContigs()}

\item{saveFile='animal_boxplot.svg'}{location in which to save boxplot file}
}
\value{
an .svg file to the destination in saveLocation
}
\description{
makeBoxPlot -- creates boxplot of contigs included in the analysis vs those excluded
}

