% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChrTable.R
\name{makeChrTable}
\alias{makeChrTable}
\title{makeChrTable -- Pulls out chromosome and length data from the header of a bam file}
\usage{
makeChrTable(bamFile, verbose = TRUE, asBed = FALSE, asRownames = TRUE)
}
\arguments{
\item{bamFile}{string of location of a bam file to extract header data from}

\item{asBed=FALSE}{if TRUE, a bed format table will be generated (chr, start, end)}

\item{verbose=TRUE}{if FALSE, no messages appear on terminal}
}
\description{
makeChrTable -- Pulls out chromosome and length data from the header of a bam file
}
\details{
makeChrTable creates a table with chromosome name and chromosome length by extracting 
header data from the supplied bam file.
}

