% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGapFromOverlap.R
\name{mapGapFromOverlap}
\alias{mapGapFromOverlap}
\title{mapGapFromOverlap -- function to co-localize strand state changes with assembly gaps}
\usage{
mapGapFromOverlap(sceFile, gapFile, verbose = TRUE, overlapNum = 4)
}
\arguments{
\item{sceFile}{data.frame of strand state change locations in BED format}

\item{gapFile}{data.frame of assembly gaps in BED format (can be downloaded from UCSC table browser)}

\item{verbose}{prints messages to the terminal (default is TRUE)}

\item{overlapNum}{Minimal number of strand state changes that overlap with a gap before assembly is cut at that location}
}
\value{
a list containing two matrices: a StrandFreqMatrix of W:C read frequencies, and a StrandReadMatrix of read counts
}
\description{
mapGapFromOverlap -- function to co-localize strand state changes with assembly gaps
}
\examples{
#Get a list of BAM files containing libraries for cells from the same organism, aligned to the same genome
#In this case these are the example BAM files provided with the package (hence the call to system.file);

example.dir <- file.path(system.file(package='contiBAIT'), 'extdata')
bam.files <- dir(example.dir, full.names=TRUE)

strand.freq <- strandSeqFreqTable(bam.files, pairedEnd = FALSE)

show(strand.freq[[1]])
show(strand.freq[[2]])

}

