% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFlankedLGs.R
\docType{methods}
\name{mergeFlankedLGs,LinkageGroupList,StrandStateMatrix-method}
\alias{mergeFlankedLGs}
\alias{mergeFlankedLGs,LinkageGroupList,StrandStateMatrix-method}
\title{mergeFlankedLGs -- searches for similarities at the ends of ordered linkage groups to chain groups together}
\usage{

  \S4method{mergeFlankedLGs}{LinkageGroupList,StrandStateMatrix}(linkageGroupList,
  strandStateMatrix, buildConsensus = 1, cluster = NULL,
  clusterParam = NULL, verbose = TRUE)
}
\arguments{
\item{linkageGroupList}{List of ordered vectors containing names of contigs belonging to each LG, of type LinkageGroupList}

\item{strandStateMatrix}{Table of type strandStateMatrix encompassing strand state for all contigs. Product of StrandSeqFreqTable.}

\item{buildConsensus}{number of contigs to take at the end of the linkage group to build a consensus strand state. Default is 5}

\item{cluster}{Number of times to recluster and take the consensus of. If NULL, clustering is 
run only once.}

\item{clusterParam}{optional \code{BiocParallelParam} specifying cluster to use for parallel execution. When NULL, execution will be serial.}

\item{verbose}{Outputs information to the terminal. Default is TRUE.}
}
\value{
a list containing a revised LinkageGroupList with merged groups, if appropriate, and 
a StrandStateMatrix with contigs reoriented, if newly merged groups were in opposite orientations.
}
\description{
mergeFlankedLGs -- searches for similarities at the ends of ordered linkage groups to chain groups together
}

