% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeLinkageGroups.R
\docType{methods}
\name{mergeLinkageGroups,LinkageGroupList,StrandStateMatrix-method}
\alias{mergeLinkageGroups}
\alias{mergeLinkageGroups,LinkageGroupList,LinkageGroupList-method}
\alias{mergeLinkageGroups,LinkageGroupList,StrandStateMatrix-method}
\title{mergeLinkageGroups -- merge very similar linkage groups, including those in reverse orientation}
\usage{
\S4method{mergeLinkageGroups}{LinkageGroupList,StrandStateMatrix}(object,
  allStrands, similarityCutoff = 0.7)
}
\arguments{
\item{object}{LinkageGroupList}

\item{allStrands}{StrandStateMatrix for all linkageGroups (usually reoriented by reorientStrandTable)}

\item{similarityCutoff}{merge contigs that are more similar this this}
}
\value{
list of indices within the allStrands matrix indicating linkage group membership,
 vector of orientation changes
}
\description{
mergeLinkageGroups -- merge very similar linkage groups, including those in reverse orientation
}

