% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLinkageGroups.R
\docType{methods}
\name{mergeLinkageGroups,LinkageGroupList,StrandStateMatrix-method}
\alias{mergeLinkageGroups}
\alias{mergeLinkageGroups,LinkageGroupList,StrandStateMatrix-method}
\alias{mergeLinkageGroups-LinkageGroupList-StrandStateMatrix-method}
\title{mergeLinkageGroups -- merge very similar linkage groups, including those in reverse orientation}
\usage{
\S4method{mergeLinkageGroups}{LinkageGroupList,StrandStateMatrix}(object,
  allStrands, clusterParam = NULL, cluster = 1, similarityCutoff = 0.7)
}
\arguments{
\item{object}{LinkageGroupList}

\item{allStrands}{StrandStateMatrix for all linkageGroups (usually reoriented by reorientStrandTable)}

\item{clusterParam}{optional \code{BiocParallelParam} specifying cluster to use for parallel execution.
When \code{NULL}, execution will be serial.}

\item{cluster}{Integer denoting the number of reclusterings to be performed for creating linkage groups (default is 1)}

\item{similarityCutoff}{merge contigs that are more similar this this}
}
\value{
list of indices within the allStrands matrix indicating linkage group membership,  
a list consisting of a strandStateMatrix (a reoriented version of allStrands), 
and a data.frame of type OrientationFrame containing contig names and orientations, as '+' or '-'.
}
\description{
mergeLinkageGroups -- merge very similar linkage groups, including those in reverse orientation
}
\examples{
data(exampleWCMatrix)
clusteredContigs <- clusterContigs(exampleWCMatrix, randomise=FALSE)

reorientedMatrix <- reorientLinkageGroups(clusteredContigs,
										  exampleWCMatrix)

exampleLGList <- mergeLinkageGroups(clusteredContigs,
									reorientedMatrix[[1]])
}

