% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderAllLinkageGroups.R
\name{orderAllLinkageGroups}
\alias{orderAllLinkageGroups}
\title{Function to call contig ordering algorithms iteratively across each linkage group element}
\usage{
orderAllLinkageGroups(linkageGroupList, strandStateMatrix, strandFreqMatrix,
  saveOrderedPDF = FALSE, orderCall = "greedy", verbose = TRUE)
}
\arguments{
\item{linkageGroupList}{list of vectors, each specifying which contigs belong in which linkage group (product of clusterContigs)}

\item{strandStateMatrix}{table of strand calls for all contigs (product of preprocessStrandTable)}

\item{strandFreqMatrix}{list with table of W:C read proportions (used for QC) and read counts (product of strandSeqFreqTable)}

\item{saveOrderedPDF}{Will return a pdf of heatmaps for each linkage group; String entered becomes the fileName (default is saveOrderedPDF=FALSE)}

\item{orderCall}{currently either 'greedy' for greedy algorithm or 'TSP' for travelling salesperson alogrithm (default is 'greedy')}

\item{verbose}{Pringts messages to the terminal. Default is TRUE}
}
\value{
a data.frame of ordered contigs with linkage group names
}
\description{
Function to call contig ordering algorithms iteratively across each linkage group element
}

