% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderContigsGreedy.R
\name{orderContigsGreedy}
\alias{orderContigsGreedy}
\title{Function to order contigs within a single linkage group using a greedy algorithms
Attempt to order contigs within}
\usage{
orderContigsGreedy(linkageGroups, allStrands, readTable, lg,
  randomAttempts = 75, verbose = TRUE)
}
\arguments{
\item{linkageGroups}{list of vector of contig names for all linkage groups (product of clusterContigs)}

\item{allStrands}{table of strand calls for all contigs}

\item{readTable}{list with table of W:C read proportions (used for QC) and read counts (product of strandSeqFreqTable)}

\item{lg}{iterger specifying the element of linkageGroups to be used (i.e. specific linkage group to try to order)}

\item{randomAttempts}{iterger specifying number of randomized clusterings to identify the best ordering. Default is 75}
}
\value{
list of two members: 1) contig names in order, 2) contigs that were combined due to being zero-distance from each other
}
\description{
Function to order contigs within a single linkage group using a greedy algorithms
Attempt to order contigs within
}

