% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OrientationFrame-class}
\alias{OrientationFrame-class}
\title{A class for storing contig orientations}
\description{
\describe{
 This class is data.frame of two character vectors that represent the orientation of contigs. 
 The first element of this data.frame is the contigs name
 The second element is the orinetation (as either + or -.
}
}

