% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherWriteBed.R
\name{otherWriteBed}
\alias{otherWriteBed}
\title{otherWriteBed -- function to write contig order to bed file}
\usage{
otherWriteBed(fileName, contigOrder, libWeight)
}
\arguments{
\item{fileName}{bed file name to write}

\item{contigOrder}{a contig ordering}

\item{ligWeight}{average quality across all libraries for a contig}
}
\value{
void -bed file written to working directory names by fileName, chromosme and state.
}
\description{
otherWriteBed -- function to write contig order to bed file
}

