% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotContigOrder.R
\name{plotContigOrder}
\alias{plotContigOrder}
\title{Plot ordering of contigs within a single linkage group.}
\usage{
plotContigOrder(contigOrder)
}
\arguments{
\item{contigOrder}{vector with the names of the contigs to plot}
}
\description{
Plot ordering of contigs within a single linkage group.
}

