% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLGDistances.R
\docType{methods}
\name{plotLGDistances,LinkageGroupList,StrandStateMatrix-method}
\alias{plotLGDistances}
\alias{plotLGDistances,LinkageGroupList,LinkageGroupList-method}
\alias{plotLGDistances,LinkageGroupList,StrandStateMatrix-method}
\title{plotLGDistances -- plots a heatmap of the distances between linkage groups}
\usage{
\S4method{plotLGDistances}{LinkageGroupList,StrandStateMatrix}(object,
  allStrands, lg = "all", labels = TRUE, state = "all",
  alreadyOrdered = FALSE)
}
\arguments{
\item{object}{LinkageGroupList}

\item{allStrands}{StrandStateMatrix for all linkageGroups (usually reoriented by reorientStrandTable)}

\item{lg}{='all' vector of integers to determine which linkage group(s) to plot. 'all' will calculate consensus
strand calls for all linkage groups and plot them side by side (default it 'all')}

\item{labels}{=TRUE if TRUE, contig names will be plotted on the axes}

\item{state}{string denoting whether only homozygous states should be used ('homo'; just WW vs CC), if a comparison should be made between
homozygous and heterozygous states only ('hetero'; just homo vs hetero), or if all three states ('all'; WW vs WC vs CC). Default is 'all'}

\item{alreadyOrdered}{if TRUE, the function will assume that the linkageGroupList is already ordered and not 
create a dendrogram. Default is FALSE}

\item{...}{additional parameters to pass to heatmap.2}
}
\value{
a heatplot of linkage group calls
}
\description{
plotLGDistances -- plots a heatmap of the distances between linkage groups
}
\examples{
data("exampleLGList")
data("exampleWCMatrix")

plotLGDistances(exampleLGList, exampleWCMatrix)
plotLGDistances(exampleLGList, exampleWCMatrix, lg=1)
}

