% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLinkageGroup.R
\name{plotLinkageGroup}
\alias{plotLinkageGroup}
\title{Plot a heatmap of contigs within a single linkage group.}
\usage{
plotLinkageGroup(groupMembers, allStrands, lgChr = NULL, labels = TRUE, ...)
}
\arguments{
\item{groupMembers}{vector with the names of the contigs to plot}

\item{allStrands}{matrix of all contigs with strand inheritance for all libraries}

\item{lgChr}{=NULL vector of true calls for chromosomes of contigs, will be plotted as side colours if 
not NULL}

\item{labels}{=TRUE if TRUE, contig names will be plotted on the axes}

\item{...}{additional parameters to pass to heatmap.2}
}
\description{
Plot a heatmap of contigs within a single linkage group.
}
\examples{
data("exampleLGList")
data("exampleWCMatrix")
plotLinkageGroup(exampleLGList[[1]], exampleWCMatrix)
}

