% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RawReadStrands-class}
\alias{RawReadStrands-class}
\title{A class for storing a data frame of directional reads from a single contig
of a single library}
\description{
\describe{
 The strand information stored in this object represent every read meeting the threshold 
 critera from the specified contig for a given library.
}
}

