% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readStrandCountsFromBAM.R
\name{readStrandCountsFromBAM}
\alias{readStrandCountsFromBAM}
\title{countDirectionReads -- function to process bam files for contiBAIT}
\usage{
readStrandCountsFromBAM(bamDir, qualLimit = 10, readLimit = 10,
  dups = FALSE, verbose = TRUE, fieldSep = ".", field = 1)
}
\arguments{
\item{bamDir}{directory to scan for bam files}

\item{qualLimit}{minimum quality for an individual read to be accepted. default is 10}

\item{readLimit}{minimum number of reads to make a state call. contigs with less reads than this limit return NA. Default is 10}

\item{dups}{remove duplicates in output file. Default is FALSE}

\item{verbose}{prints messages to the terminal (default is TRUE)}

\item{fieldSep}{The seperator to denote different fields in a fileName. Default is '.'}

\item{field}{The region to use as a file identifier eg. 401.sort.bam has 3 fields seperated by periods. default is 1.}
}
\value{
list of two data.frames. The first displays state calls for every contig (rows) in every library (cols). The second is a matching data.frame with the number of reads used to make the call
}
\description{
countDirectionReads -- function to process bam files for contiBAIT
}

