% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorientAndMergeLGs.R
\docType{methods}
\name{reorientAndMergeLGs,LinkageGroupList,StrandStateMatrix-method}
\alias{reorientAndMergeLGs,LinkageGroupList,StrandStateMatrix-method}
\alias{reorientAndMergeLGs}
\alias{reorientAndMergeLGs-LinkageGroupList-StrandStateMatrix-method}
\title{reorientAndMergeLGs uses a simple dissimilarity to find misoriented fragments within linkage groups.}
\usage{
\S4method{reorientAndMergeLGs}{LinkageGroupList,StrandStateMatrix}(object,
  allStrands, cluster = NULL, clusterParam = NULL, similarityCutoff = 0.9,
  verbose = TRUE)
}
\arguments{
\item{object}{List of vectors containing names of contigs belonging to each LG.}

\item{allStrands}{Table of type \code{strandStateMatrix} encompassing strand state for all contigs. Product of StrandSeqFreqTable.}

\item{cluster}{Number of times to recluster and take the consensus of. If NULL, clustering is 
run only once.}

\item{clusterParam}{optional \code{BiocParallelParam} specifying cluster to use for parallel execution.
When \code{NULL}, execution will be serial.}

\item{similarityCutoff}{merge contigs that are more similar this this}

\item{verbose}{Outputs information to the terminal. Default is TRUE.}
}
\value{
a list consisting of a strandStateMatrix (a reoriented version of allStrands), a ChrTable 
containing contig names and orientations, as '+' or '-' and a merged LinkageGroupList.
}
\description{
reorientAndMergeLGs uses a simple dissimilarity to find misoriented fragments within linkage groups.
}
\examples{
data(exampleLGList)
data(exampleWCMatrix)

reorientedMatrix <- reorientAndMergeLGs(exampleLGList,
										  exampleWCMatrix)

# Note that in this example data, everything is correctly oreiented to
# to begin with, so all contigs come out as + orientation
}
