% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorientLinkageGroups.R
\docType{methods}
\name{reorientLinkageGroups,LinkageGroupList,StrandStateMatrix-method}
\alias{reorientLinkageGroups}
\alias{reorientLinkageGroups,LinkageGroupList,LinkageGroupList-method,}
\alias{reorientLinkageGroups,LinkageGroupList,StrandStateMatrix-method}
\alias{strandStateMatrix,}
\alias{strandStateMatrix-method}
\title{reorientLinkageGroups uses a simple dissimilarity to find misoriented fragments within linkage groups.}
\usage{
\S4method{reorientLinkageGroups}{LinkageGroupList,StrandStateMatrix}(object,
  allStrands, verbose = TRUE)
}
\arguments{
\item{object}{List of vectors containing names of contigs belonging to each LG.}

\item{allStrands}{Table of strand state for all contigs. Product of StrandSeqFreqTable.}

\item{verbose}{Outputs information to the terminal. Default is TRUE.}
}
\value{
a list consisting of a strandStateMatrix (a reoriented version of allStrands), and a data.frame of contig names and orientations, as '+' or '-'.
}
\description{
reorientLinkageGroups uses a simple dissimilarity to find misoriented fragments within linkage groups.
}

