% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reorientLinkageGroups.R
\name{reorientLinkageGroups}
\alias{reorientLinkageGroups}
\title{reorientLinkageGroups uses a greedy algorithm to try to find the correct orientations of linkage groups}
\usage{
reorientLinkageGroups(linkageGroups, allStrands, dissimilarityCutoff = 0.3,
  maxiter = 100)
}
\arguments{
\item{linkageGroups}{List of vectors containing names of contigs belonging to each LG}

\item{allStrands}{Table of strand state for all contigs.}

\item{dissimilarityCutoff}{Try to reorient all LGs more dissimilar than this to any other LG}

\item{maxiter}{Maximum number of iterations to try reoriented for (to prevent possible infinite loops)}
}
\value{
a vector of orientations, as '+' or '-', in the order of linkageGroups
}
\description{
reorientLinkageGroups uses a greedy algorithm to try to find the correct orientations of linkage groups
}

