% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,ChrTable-method}
\alias{show,ChrTable-method}
\title{show-methods}
\usage{
\S4method{show}{ChrTable}(object)
}
\arguments{
\item{object}{a ChrTable}
}
\description{
Shows a ChrTable
}

