% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,ContigOrdering-method}
\alias{show,ContigOrdering-method}
\title{show-methods}
\usage{
\S4method{show}{ContigOrdering}(object)
}
\arguments{
\item{object}{a ContigOrdering}
}
\value{
nothing
}
\description{
Shows a ContigOrdering
}
