% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,LinkageGroupList-method}
\alias{show,LinkageGroupList-method}
\title{show-methods}
\usage{
\S4method{show}{LinkageGroupList}(object)
}
\arguments{
\item{object}{a LinkageGrouplist}
}
\description{
Shows a LinkageGroupList
}

