% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,OrientationFrame-method}
\alias{show,OrientationFrame-method}
\title{show-methods}
\usage{
\S4method{show}{OrientationFrame}(object)
}
\arguments{
\item{object}{a OrientationFrame}
}
\value{
nothing
}
\description{
Shows a OrientationFrame
}

