% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,RawReadStrands-method}
\alias{show,RawReadStrands-method}
\title{show-methods}
\usage{
\S4method{show}{RawReadStrands}(object)
}
\arguments{
\item{object}{a RawReadStrands}
}
\value{
nothing
}
\description{
Shows a RawReadStrands
}

