% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,StrandReadMatrix-method}
\alias{show,StrandReadMatrix-method}
\title{show-methods}
\usage{
\S4method{show}{StrandReadMatrix}(object)
}
\arguments{
\item{object}{a StrandReadMatrix}
}
\value{
nothing
}
\description{
Shows a StrandReadMatrix
}

