% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,StrandStateList-method}
\alias{show,StrandStateList-method}
\title{show-methods}
\usage{
\S4method{show}{StrandStateList}(object)
}
\arguments{
\item{object}{a StrandStateList}
}
\value{
nothing
}
\description{
Shows a StrandStateList
}
