% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,StrandStateMatrix-method}
\alias{show,StrandStateMatrix-method}
\title{show-methods}
\usage{
\S4method{show}{StrandStateMatrix}(object)
}
\arguments{
\item{object}{a StrandStateMatrix}
}
\description{
show-methods
}

