% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{StrandFreqMatrix-class}
\alias{StrandFreqMatrix-class}
\alias{StrandFreqMatrix}
\title{A class for storing a matrix of frequencies of Watson to Crick reads
for a set of contigs over several libraries}
\usage{
StrandFreqMatrix(counts = matrix(double()))
}
\arguments{
\item{counts}{a double matrix of read count ratios}
}
\value{
a \code{StrandFreqMatrix}
}
\description{
\describe{
 The strand information stored in this object is the ratio of Watson to Crick reads
 mapping to each contig in each library (cell). This should fall within the range (-1,1).
This class simply extends matrix, but with additional validity checking.
}
}
\examples{
data("exampleWatsonFreq")
data("exampleCrickFreq")
frequencyMatrix <- sapply(1:ncol(exampleCrickFreq), 
function(colNum){exampleCrickFreq[,colNum] / exampleWatsonFreq[,colNum]})

StrandFreqMatrix(frequencyMatrix)
}
