% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{StrandMatrix-class}
\alias{StrandMatrix-class}
\title{A class for storing a matrix of read count-based strand states 
of a set of contigs over several libraries}
\description{
\describe{
 The strand information stored in this object is the ratio of Watson to Crick reads
 mapping to each contig in each library (cell). This should fall within the range (-1,1).
This class simply extends matrix, but with additional validity checking.
}
}

