% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{StrandReadMatrix-class}
\alias{StrandReadMatrix-class}
\title{A class for storing read counts for a set of contigs over several libraries}
\description{
\describe{
 The information stored in this class is simple read counts, so should be integers >=0.
}
}

