% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandSeqFreqTable.R
\name{strandSeqFreqTable}
\alias{strandSeqFreqTable}
\title{strandSeqFreqTable -- function to process bam files for contiBAIT}
\usage{
strandSeqFreqTable(bamFileList, fieldSep = ".", field = 1, qual = 0,
  rmdup = TRUE, verbose = TRUE, filter = FALSE, tileChunk = 1e+05,
  pairedEnd = TRUE)
}
\arguments{
\item{bamFileList}{vector containing the location of the bams file to be read}

\item{fieldSep}{The field seperator of the bam file to use to define the field. Default is '.'}

\item{qual}{Mapping quality threshold. Default is 0}

\item{rmdup}{remove duplicates in output file. Default is TRUE}

\item{verbose}{prints messages to the terminal (default is TRUE)}

\item{filter}{additional file to split chromosomes based on locations. If this parameter is blank,
a filter table will be automatically generated from the header of the first file in bamFileList}

\item{tileChunk}{Number of reads to split bam files into (smaller number requires less RAM). Default is 100000.}

\item{pairedEnd}{Whether the bam files being read are in paired end format. Default is TRUE. Note,
since paired reads will be the same direction, only first mate read of pair is used in output}

\item{filed}{The field of the bam file name to use as an index (default is 1)}
}
\value{
a list containing two matrices: a StrandFreqMatrix of W:C read frequencies, and a StrandReadMatrix of read counts
}
\description{
strandSeqFreqTable -- function to process bam files for contiBAIT
}
\examples{
bam.files <- sapply(dir('extdata'), function(single.file){system.file('extdata', single.file, package='contiBAIT')})
strand.freq <- strandSeqFreqTable(bam.files, pairedEnd = FALSE)

show(strand.freq[[1]])
show(strand.freq[[2]])
}

