% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{StrandStateList-class}
\alias{StrandStateList-class}
\alias{StrandStateList}
\title{A class for storing StrandStateList lists for contigs}
\usage{
StrandStateList(strandGroupList = list(), names = character())
}
\arguments{
\item{strandGroupList}{a list of StrandStateMatrix elements, with each primary element representing a StrandStateMatrix containing ordered contigs from a LinkageGroupList element}

\item{names}{a vector of names of StrandStateMatrix elements}
}
\value{
a \code{StrandStateList}
}
\description{
\describe{
 This class is a list of StrandStateMatrices, each a subset of the StrandStateMatrix split by linkagegroups
}
}
