% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{StrandStateMatrix-class}
\alias{StrandStateMatrix-class}
\title{A class for storing a data frame of discrete strand states
of a set of contigs over several libraries}
\description{
\describe{
 The strand information stored in this object is a call of the strand state
 of each contig in each library.
 mapping to each contig in each library (cell). This should fall within the range (-1,1).
	This class simply extends matrix, but with additional validity checking.
}
}

