% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.bed.R
\name{write.bed}
\alias{write.bed}
\title{write.bed -- function to merge and write like-data to a bed file}
\usage{
write.bed(libList, chr, libCall = 10, splitBed = FALSE, skip = 3,
  verbose = TRUE, byChr = TRUE, file = "bedfile", path = ".")
}
\arguments{
\item{libList}{vector of library names to be analyzed; created using mergeLibraries}

\item{chr}{chromoosome name to be analyzed}

\item{libCall}{Number reflecting the state of the bed files; 1=Crick, 3=Watson, 10=analyze allFiles. Default is 10}

\item{splitBed}{Logical value which when TRUE will place bed headers within file, generating split (unmerged) bedfiles. Default is FALSE}

\item{skip}{File containing known gap locations in bed format. Either gapFile or binData must be specified}

\item{verbose}{prints messages to the terminal (default is TRUE)}

\item{byChr}{Creates bedFile consisting of reads from only the chromosome being analyzed. Default is TRUE}

\item{file}{outPut fileName. Default is 'bedfile'}

\item{path}{the location of the bed files in which to merge. Default is current directory.}
}
\value{
bed file printed to working directory names by fileName, chromosme and state.
}
\description{
write.bed -- function to merge and write like-data to a bed file
}

