% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBed.R
\docType{methods}
\name{writeBed,ChrTable,OrientationFrame,ContigOrdering-method}
\alias{writeBed}
\alias{writeBed,ChrTable,ChrTable-method,OrientationFrame,OrientationFrame-method,ContigOrdering,ContigOrdering-method}
\alias{writeBed,ChrTable,OrientationFrame,ContigOrdering-method}
\title{function to write contig order to BED file}
\usage{
\S4method{writeBed}{ChrTable,OrientationFrame,ContigOrdering}(chrTable,
  orientationData, contigOrder, libWeight = NULL,
  fileName = "contiBAIT_assembly")
}
\arguments{
\item{chrTable}{an object of type ChrTable with rownames matching contig names. Product of makeChrTable}

\item{orientationData}{data.frame of contig and strand (with rownames matching contig names). Product of reorientLinkageGroups[[2]]}

\item{contigOrder}{an object of type ContigOrdering with ordered Linkage Groups and contigs. Product of orderAllLinkageGroups}

\item{libWeight}{average quality across all libraries for a contig}

\item{fileName}{character string for bed file name to write}
}
\value{
void; BED file written to working directory as fileName.bed.
}
\description{
function to write contig order to BED file
}

