% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ChrTable-class}
\alias{ChrTable-class}
\alias{ChrTable}
\title{A class for storing chromosome/fragment lengths}
\usage{
ChrTable(chrRanges = GRanges())
}
\arguments{
\item{chrRanges}{a GRanges object with a meta column 
called name, which represents the fragment name}
}
\value{
a \code{ChrTable}
}
\description{
\describe{
 This class is a GRanges object with a meta column called name, which represents the fragment name.
}
}
