% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ContigOrdering-class}
\alias{ContigOrdering}
\alias{ContigOrdering-class}
\title{A class for storing contig ordering of a linkage group}
\usage{
ContigOrdering(ordering = character())
}
\arguments{
\item{ordering}{a matrix of two character vectors that represent the calculated ordering of a linkage group. 
The first element of this matrix is the Linkage Group sub-setted by contigs with equal strand states across all libraries
in the calculated order. 
The second element is the names of names of each contig in the calculated order.}
}
\value{
a \code{ContigOrdering}
}
\description{
\describe{
 This class is a matrix of two character vectors that represent the calculated ordering of a linkage group. 
 The first element of this matrix is the Linkage Group sub-setted by contigs with equal strand states across all libraries
 in the calculated order. 
 The second element is the names of names of each contig in the calculated order.
}

Constructor for ContigOrdering
}
\examples{
thisOrdering <- ContigOrdering(matrix(ncol=2, c( "LG1.11", "chr2:1000820-2001640", 
			"LG1.1", "chr2:3002461-4003281")))
}

