% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ContigOrdering-class}
\alias{ContigOrdering-class}
\title{A class for storing contig ordering of a linkage group}
\description{
\describe{
 This class is data.frame of two character vectors that represent the calculated ordering of a linkage group. 
 The first element of this data.frame is the Linkage Group sub-setted by contigs with equal strand states across all libraries
 in the calculated order. 
 The second element is the names of names of each contig in the calculated order.
}
}

