% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleDividedChr}
\alias{exampleDividedChr}
\title{Example of a divided chromosome, containing contigs and their lengths}
\description{
Example of a divided chromosome, containing contigs and their lengths
}
\keyword{data}
