% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{exampleWWCCMatrix}
\alias{exampleWWCCMatrix}
\title{An example StrandStateMatrix containing only WW and CC calls for contigs}
\description{
An example StrandStateMatrix containing only WW and CC calls for contigs
}
\keyword{data}

