% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogramPlot.R
\docType{methods}
\name{ideogramPlot,StrandReadMatrix,StrandReadMatrix,ChrTable-method}
\alias{ChrTable,}
\alias{ChrTable-method}
\alias{ideogramPlot}
\alias{ideogramPlot,StrandReadMatrix,StrandReadMatrix,ChrTable-method}
\alias{ideogramPlot,StrandReadMatrix,StrandReadMatrix-method,}
\title{ideogramPlot -- plots BAIT-like ideograms}
\usage{

  \S4method{ideogramPlot}{StrandReadMatrix,StrandReadMatrix,ChrTable}(WatsonFreqList,
  CrickFreqList, chrTable, plotBy = "lib", showPage = FALSE,
  orderFrame = FALSE, orientationData = NULL, verbose = TRUE)
}
\arguments{
\item{WatsonFreqList}{data.frame of Watson calls. Product of strandSeqFreqTable[[3]] when BAITtables=TRUE}

\item{CrickFreqList}{data.frame of Crick calls. Product of strandSeqFreqTable[[4]] when BAITtables=TRUE}

\item{chrTable}{A data.frame consisting of chromosomes and lengths. Generated by makeChrTable(). Note rownames equal to chromosome names are required}

\item{plotBy}{Whether to generate a plot for each library ('lib') or a plot for each chromosome ('chr')}

\item{showPage}{Integer specifying which page to plot if plotBy='chr' selected. Useful when not plotting to a file. Default is FALSE}

\item{orderFrame}{ordered data.frame of contigs (produced by orderAllLinkageGroups). Default is FALSE, where plots will be made from elements in chrTable.}

\item{orientationData}{data.frame of contig orientations of type OrientationFrame telling which reads to flip Watson and Crick counts}

\item{verbose}{prints messages to the terminal (default is TRUE)}
}
\value{
ordered contigs in bed format. Depending on options, intermediate files and plots will also be generated
}
\description{
ideogramPlot -- plots BAIT-like ideograms
}
\examples{
data("exampleWatsonFreq")
data("exampleCrickFreq")
data('exampleDividedChr')

singleWatsonLibrary <- new('StrandReadMatrix', exampleWatsonFreq[,2, drop=FALSE])
singleCrickLibrary <- new('StrandReadMatrix', exampleCrickFreq[,2, drop=FALSE]) 

ideogramPlot(singleWatsonLibrary, singleCrickLibrary, exampleDividedChr)
}

