% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LinkageGroupList-class}
\alias{LinkageGroupList-class}
\title{A class for storing linkage group calls for contigs}
\description{
\describe{
 This class is simply a list of character strings containing the names of linkage groups. 
}
}

