% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateMisorients.R
\docType{methods}
\name{locateMisorients,GRanges-method}
\alias{locateMisorients,GRanges-method}
\alias{locateMisorients}
\alias{locateMisorients,locateMisorients-GRanges-method}
\title{locateMisorients -- function to identify libraries that hare similar WC patterns on chromosomes}
\usage{
\S4method{locateMisorients}{GRanges}(compiledGrange, gapFile = NULL,
  stateNum = 3, readCutOff = 40, verbose = TRUE)
}
\arguments{
\item{compiledGrange}{A GRanges object consisting of read locations. Can be an individual file or the product of thoroughBed}

\item{gapFile}{A GRanges object consisting of start and end locations of assembly gaps (defaul it NULL)}

\item{stateNum}{The number of expected strand states. Default is 3 (WW, WC and CC). Function may exhibit unusual behaviour is changed}

\item{readCutOff}{The minimal number of reads required to make an accurate strand state call. Default is 40.}

\item{verbose}{prints messages to the terminal (default is TRUE)}
}
\value{
a list of ChrTable objects. The first is a ChrTable of misorientations detected. The second is a ChrTable of chimera detected.
Output can be used in downstream functions (strandSeqFreqTable)
}
\description{
locateMisorients -- function to identify libraries that hare similar WC patterns on chromosomes
}
