% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBoxPlot.R
\docType{methods}
\name{makeBoxPlot,ChrTable,LinkageGroupList-method}
\alias{makeBoxPlot}
\alias{makeBoxPlot,ChrTable,LinkageGroupList-method}
\alias{makeBoxPlot-ChrTable-LinkageGroupList-method}
\title{makeBoxPlot -- creates boxplot of contigs included in the analysis vs those excluded}
\usage{
\S4method{makeBoxPlot}{ChrTable,LinkageGroupList}(chrTable, linkage.contigs)
}
\arguments{
\item{chrTable}{A data.frame consisting of chromosomes and lengths. Generated by makeChrTable(). Note rownames equal to chromosome names are required}

\item{linkage.contigs}{A list of clustered contigs, generated by clusterContigs()}
}
\value{
a box plot of included and excluded contigs
}
\description{
makeBoxPlot -- creates boxplot of contigs included in the analysis vs those excluded
}
\examples{
#make an example barplot of data

data("exampleLGList")
data("exampleChrTable")
makeBoxPlot(exampleChrTable, exampleLGList)

}

