% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGapFromOverlap.R
\name{mapGapFromOverlap}
\alias{mapGapFromOverlap}
\title{mapGapFromOverlap -- function to co-localize strand state changes with assembly gaps}
\usage{
mapGapFromOverlap(sceFile, gapFile, chrTable, verbose = TRUE,
  overlapNum = 4)
}
\arguments{
\item{sceFile}{GRanges object of strand state change locations in BED format}

\item{gapFile}{GRanges object of assembly gaps in BED format (can be downloaded from UCSC table browser)}

\item{chrTable}{GRanges object of chromosome table (product of makeChrTable)}

\item{verbose}{prints messages to the terminal (default is TRUE)}

\item{overlapNum}{Minimal number of strand state changes that overlap with a gap before assembly is cut at that location}
}
\value{
a GRanges object of all contigs split by regions where the sceFile and gapFile GRanges objects overlap.
}
\description{
mapGapFromOverlap -- function to co-localize strand state changes with assembly gaps
}

