% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderContigsGreedy.R
\name{orderContigsGreedy}
\alias{orderContigsGreedy}
\title{Function to order contigs within a single linkage group using a greedy algorithms
Attempt to order contigs within}
\usage{
orderContigsGreedy(linkageGroupReadTable, randomAttempts = 75,
  nProcesses = 1, verbose = TRUE)
}
\arguments{
\item{linkageGroupReadTable}{dataframe of strand calls (product of combineZeroDists or preprocessStrandTable)}

\item{randomAttempts}{number of times to repeat the greedy algortihm with a random restart}

\item{nProcesses}{number of processes to attempt ordering in parallel}

\item{verbose}{whether to print verbose messages}
}
\value{
list of two members: 1) contig names in order, 2) the original data.frame entered into function correctly ordered
}
\description{
Function to order contigs within a single linkage group using a greedy algorithms
Attempt to order contigs within
}

