% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderContigsTSP.R
\name{orderContigsTSP}
\alias{orderContigsTSP}
\title{Attempt to order contigs within linkage groups using travelling salesperson algorithm}
\usage{
orderContigsTSP(linkageGroupReadTable, reps)
}
\arguments{
\item{linkageGroupReadTable}{dataframe of strand calls (product of combineZeroDists or preprocessStrandTable)}

\item{reps}{Number of repetitions for he 2-opt TSP solver algorithm}
}
\value{
list of two members: 1) contig names in order, 2) the original data.frame entered into function correctly ordered
}
\description{
Attempt to order contigs within linkage groups using travelling salesperson algorithm
}
