% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderContigsTSP.R
\name{orderContigsTSP}
\alias{orderContigsTSP}
\title{Attempt to order contigs within}
\usage{
orderContigsTSP(linkageGroupReadTable)
}
\arguments{
\item{linkageGroups}{list of vector of contig names for all linkage groups}

\item{allStrands}{table of strand calls for all contigs}

\item{readTable}{table of W:C read proportions (used for QC)}

\item{lg}{iterger specifying the element of linkageGroups to be used (i.e. specific linkage group to try to order)}
}
\value{
list of two members: 1) contig names in order, 2) contigs that were combined due to being zero-distance from each other
}
\description{
Attempt to order contigs within
}

