% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderWithinGroup.R
\name{orderWithinGroup}
\alias{orderWithinGroup}
\title{Attempt to order contigs within}
\usage{
orderWithinGroup(linkageGroup, readTable, allStrands, method = "greedy",
  contigWeight = NULL)
}
\arguments{
\item{linkageGroup}{vector of names of contigs in this linkage group to try to order}

\item{readTable}{table of W:C read proportions (used for QC)}

\item{allStrands}{table of strand calls for all contigs}

\item{method}{algorithm to order the contigs}

\item{contigWeight}{optional order in which to place contigs using greedy algorithm}

\item{lg}{numeric value defining which linkage groups}
}
\value{
list of two members: 1) contig names in order, 2) contigs that were combined due to being zero-distance from each other
}
\description{
Attempt to order contigs within
}

