% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContigOrder.R
\docType{methods}
\name{plotContigOrder,ContigOrdering-method}
\alias{plotContigOrder}
\alias{plotContigOrder,ContigOrdering-method}
\title{Plot ordering of contigs within a single linkage group.}
\usage{
\S4method{plotContigOrder}{ContigOrdering}(contigOrder, lg = "all",
  verbose = TRUE)
}
\arguments{
\item{contigOrder}{matrix from orderAllContigs with the subdivided linkage groups and the names of the contigs to plot}

\item{lg}{Integer specifying the linkage group by which to plot. Default is all}

\item{verbose}{prints messages to the terminal (default is TRUE)}
}
\value{
A ggplot object (which will be plotted automatically if not assigned).
}
\description{
Plot ordering of contigs within a single linkage group.
}
\examples{
#Get a data.frame of ordered contigs from cells from the same organism, aligned to the same genome
 
data("exampleContigOrder")

plotContigOrder(exampleContigOrder[[1]], lg=1)
}

