% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLGDistances.R
\docType{methods}
\name{plotLGDistances,LinkageGroupList,StrandStateMatrix-method}
\alias{plotLGDistances}
\alias{plotLGDistances,LinkageGroupList,LinkageGroupList-method}
\alias{plotLGDistances,LinkageGroupList,StrandStateMatrix-method}
\title{plotLGDistances -- plots a heatmap of the distances between linkage groups}
\usage{
\S4method{plotLGDistances}{LinkageGroupList,StrandStateMatrix}(object,
  allStrands)
}
\arguments{
\item{object}{LinkageGroupList}

\item{allStrands}{StrandStateMatrix for all linkageGroups (usually reoriented by reorientStrandTable)}

\item{saveFile}{string defining file name of plot. default is FALSE (no plot saved)}
}
\description{
plotLGDistances -- plots a heatmap of the distances between linkage groups
}

