% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWCdistribution.R
\docType{methods}
\name{plotWCdistribution,StrandFreqMatrix-method}
\alias{plotWCdistribution}
\alias{plotWCdistribution,StrandFreqMatrix,StrandFreqMatrix-method}
\alias{plotWCdistribution,StrandFreqMatrix-method}
\title{Creates median distribution boxplots across all libraries and contigs}
\usage{
\S4method{plotWCdistribution}{StrandFreqMatrix}(object, filterThreshold = 0.8)
}
\arguments{
\item{object}{object of class StrandFreqMatrix (product of strandSeqFreqTable)}

\item{filterThreshold}{numeric value used in assessing the threshold for homozygous strand calls. Default is 0.8.}
}
\value{
nothing, just plots.
}
\description{
Creates median distribution boxplots across all libraries and contigs
}
\examples{
data("exampleStrandFreq")

plotWCdistribution(exampleStrandFreq, filterThreshold=0.8)
}

