% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWCdistribution.R
\docType{methods}
\name{plotWCDistribution,StrandFreqMatrix-method}
\alias{plotWCDistribution}
\alias{plotWCDistribution,StrandFreqMatrix,StrandFreqMatrix-method}
\alias{plotWCDistribution,StrandFreqMatrix-method}
\title{Creates median distribution boxplots across all libraries and contigs}
\usage{
\S4method{plotWCDistribution}{StrandFreqMatrix}(object, filterThreshold = 0.8,
  saveFile = FALSE)
}
\arguments{
\item{object}{object of class StrandFreqMatrix (product of strandSeqFreqTable)}

\item{filterThreshold}{numeric value used in assessing the threshold for homozygous strand calls. Default is 0.8.}

\item{saveFile}{character denoting whether plot should be saved. Plots to R graphics when FALSE. Default is FALSE}
}
\description{
Creates median distribution boxplots across all libraries and contigs
}

