% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessStrandTable.R
\name{preprocessStrandTable}
\alias{preprocessStrandTable}
\title{preprocessStrandTable -- remove low quality libraries and contigs before attempting to build
a genome}
\usage{
preprocessStrandTable(strandTable, strandTableThreshold = 0.8,
  filterThreshold = 0.8, orderMethod = "libsAndConc",
  lowQualThreshold = 0.9, verbose = TRUE, minLib = 10,
  ignoreInternalQual = FALSE)
}
\arguments{
\item{strandTable}{data.frame containing the strand table to use as input}

\item{strandTableThreshold=0.8}{threshold at which to call a contig WW or CC rather than WC}

\item{filterThreshold=0.8}{maximum number of libraries a contig can be NA or WC in}

\item{orderMethod='libsAndConc'}{the method to oder contigs. currently libsAndConc only option. Set to FALSE to not order contigs based on library quality}

\item{lowQualThreshold=0.9}{background threshold at which to toss an entire library}

\item{ignoreInternalQual=FALSE}{logical that prevents function for making an overall assessment of library quality. Very chimeric assemblies can appear low quality across all libraries.}

\item{verbose=TRUE}{messages written to terminal}
}
\value{
A list of three matrices -- 1: WW/WC/WW; 2: WW/CC; 3: sex contigs
}
\description{
preprocessStrandTable -- remove low quality libraries and contigs before attempting to build
a genome
}

