% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reorientStrandTable.R
\docType{methods}
\name{reorientStrandTable,StrandStateMatrix-method}
\alias{reorientStrandTable}
\alias{reorientStrandTable,StrandStateMatrix,StrandStateMatrix-method}
\alias{reorientStrandTable,StrandStateMatrix-method}
\title{Reorient a strand table, given the orientation vector computed from \code{reorientLinkageGroups}}
\usage{
\S4method{reorientStrandTable}{StrandStateMatrix}(object,
  linkageGroups = NULL, orientation = NULL)
}
\arguments{
\item{object}{Strand table to reorient}

\item{linkageGroups}{List of vectors containing names of contigs belonging to each LG}

\item{orientation}{vector of linkage group orientations, as '+' or '-'}
}
\value{
the strand table, reoriented according to the linkage group orientations
}
\description{
Reorient a strand table, given the orientation vector computed from \code{reorientLinkageGroups}
}
\examples{
data(exampleWWCCMatrix)
data(exampleWCMatrix)

clusteredContigs <- clusterContigs(exampleWWCCMatrix, verbose=FALSE)
show(clusteredContigs)
show(clusteredContigs[[1]])

LGOrientations <- reorientLinkageGroups(clusteredContigs, exampleWWCCMatrix)
reorientedMatrix <- reorientStrandTable(exampleWCMatrix, linkageGroups = clusteredContigs, orientation = LGOrientations)
mergedLinkageGroups <- mergeLinkageGroups(clusteredContigs,reorientedMatrix)
}

