% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show,RawReadStrands-method}
\alias{show,RawReadStrands-method}
\title{show-methods}
\usage{
\S4method{show}{RawReadStrands}(object)
}
\arguments{
\item{object}{a RawReadStrands}
}
\description{
show-methods
}

