% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{StrandReadMatrix-class}
\alias{StrandReadMatrix-class}
\alias{StrandReadMatrix}
\title{A class for storing read counts for a set of contigs over several libraries}
\usage{
StrandReadMatrix(counts = matrix(integer()))
}
\arguments{
\item{counts}{an integer matrix of read counts}
}
\value{
a \code{StrandReadMatrix}
}
\description{
\describe{
 The information stored in this class is simple read counts, so should be integers >=0.
}
}
\examples{
data("exampleWatsonFreq")
StrandReadMatrix(exampleWatsonFreq[,2, drop=FALSE])
}
