% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBed.R
\name{writeBed}
\alias{writeBed}
\title{function to write contig order to BED file}
\usage{
writeBed(fileName, contigOrder, libWeight)
}
\arguments{
\item{fileName}{bed file name to write}

\item{contigOrder}{a contig ordering}

\item{ligWeight}{average quality across all libraries for a contig}
}
\value{
void; BED file written to working directory names by fileName, chromosome and state.
}
\description{
function to write contig order to BED file
}

