% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBed.R
\docType{methods}
\name{writeBed,ChrTable,OrientationFrame,ContigOrdering-method}
\alias{writeBed}
\alias{writeBed,ChrTable,OrientationFrame,ContigOrdering-method}
\alias{writeBed-ChrTable-OrientationFrame-ContigOrdering-method}
\title{function to write contig order to BED file}
\usage{
\S4method{writeBed}{ChrTable,OrientationFrame,ContigOrdering}(chrTable,
  orientationData, contigOrder, libWeight = NULL,
  file = "contiBAIT_assembly.bed")
}
\arguments{
\item{chrTable}{an object of type ChrTable with rownames matching contig names. Product of makeChrTable}

\item{orientationData}{data.frame of contig and strand (with rownames matching contig names). Product of reorientLinkageGroups[[2]]}

\item{contigOrder}{an object of type ContigOrdering with ordered Linkage Groups and contigs. Product of orderAllLinkageGroups}

\item{libWeight}{average quality across all libraries for a contig}

\item{file}{character string for bed file name to write}
}
\value{
NULL; BED file written to file
}
\description{
function to write contig order to BED file
}
\examples{
\dontrun{

data(exampleDividedChr)

writeBed(exampleDividedChr,
		 reorientedMatrix[[2]],
		 contigOrder)
}
}

