% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBed.R
\docType{methods}
\name{writeBed,ANY,ChrTable,ContigOrdering-method}
\alias{writeBed,ANY,ChrTable,ContigOrdering-method}
\alias{writeBed}
\alias{writeBed-ChrTable-ContigOrdering-method}
\title{function to write contig order to BED file}
\usage{
\S4method{writeBed}{ANY,ChrTable,ContigOrdering}(chrTable, orientationData,
  contigOrder, libWeight = NULL, file = "contiBAIT_assembly.bed")
}
\arguments{
\item{chrTable}{a GRanges object with a 'name' meta column matching contig names. Product of makeChrTable}

\item{orientationData}{ChrTable of contig and strand (with rownames matching contig names). Product of reorientAndMergeLGs[[2]]}

\item{contigOrder}{an object of type ContigOrdering with ordered Linkage Groups and contigs. Product of orderAllLinkageGroups}

\item{libWeight}{average quality across all libraries for a contig}

\item{file}{character string for bed file name to write}
}
\value{
NULL; BED file written to file
}
\description{
function to write contig order to BED file
}
\examples{
\dontrun{

data("exampleDividedChr")

writeBed(exampleDividedChr,
		 reorientedMatrix[[2]],
		 contigOrder)
}
}
